@@ -28,7 +28,7 @@ import MetadataGroup from "./metadata/MetadataGroup.vue";
 import { isoDuration } from "@musement/iso-duration";
 import { mapState } from "vuex";
 // Register custom fields for the metadata rendering
-// eslint-disable-next-line
+ 
 import "../../fields.config";
 
 export default {
